requirements.txt

flask
pywebview
requests
supabase
pyautogui
qrcode
pillow
firebase-admin
